requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
pillow>=8.0.0
matplotlib>=3.3.0
torch>=1.10.0
torchvision>=0.11.1 